Move Stripe actions to the Stripe module
