+++ os/server/payments/stripe/src/jvmMain/kotlin/com/wasmo/stripe/StripeActionSource.kt
@@ -1,4 +1,4 @@
-package com.wasmo.computers
+package com.wasmo.stripe
 
 import com.wasmo.framework.ActionSource
 import com.wasmo.framework.ActionSource.Binder
